+++ R/distancemap.R
@@ -10,6 +10,7 @@
 #' @importFrom pheatmap pheatmap
 #' @importFrom grDevices colorRampPalette
 #' @import RColorBrewer
+#' @import grid unit
 #' @export
 #' 
 plotDistMap <- function(obj, type = "b", size_d = NULL, size_s = 5) {
@@ -22,16 +23,27 @@ plotDistMap <- function(obj, type = "b", size_d = NULL, size_s = 5) {
       
       # Plot distance maps of domains before GraphSOM ------------------------
       if (!is.null(size_d)) {
-        pheatmap(distmat, cluster_cols = F, cluster_rows = F, 
-                 color = colorRampPalette(brewer.pal(n = 7, name = "Blues"))(100), na_col = "black",
-                 main = "Distance of domains", 
-                 fontsize_row = size_d, fontsize_col = size_d,
-                 cellheight = size_d, cellwidth = size_d
-                 )
+        p <- pheatmap(distmat, cluster_cols = F, cluster_rows = F, 
+                      color = colorRampPalette(brewer.pal(n = 7, name = "Blues"))(max(eobj@dist[eobj@dist != Inf]) + 1), na_col = "black",
+                      main = "Distance of domains", 
+                      fontsize_row = size_d, fontsize_col = size_d,
+                      cellheight = size_d, cellwidth = size_d, silent = T,
+                      legend_breaks = 0:max(eobj@dist[eobj@dist != Inf]),
+                      legend_labels = 0:max(eobj@dist[eobj@dist != Inf])
+                      )
+        p <- .modLabelPositions(p)
+        plot.new()
+        print(p)
       } else {
-        pheatmap(distmat, cluster_cols = F, cluster_rows = F, 
-                 color = colorRampPalette(brewer.pal(n = 7, name = "Blues"))(100), na_col = "black",
-                 main = "Distance of domains")
+        p <- pheatmap(distmat, cluster_cols = F, cluster_rows = F, 
+                      color = colorRampPalette(brewer.pal(n = 7, name = "Blues"))(max(eobj@dist[eobj@dist != Inf]) + 1), na_col = "black",
+                      main = "Distance of domains", silent = T, 
+                      legend_breaks = 0:max(eobj@dist[eobj@dist != Inf]),
+                      legend_labels = 0:max(eobj@dist[eobj@dist != Inf])
+                      )
+        p <- .modLabelPositions(p)
+        plot.new()
+        print(p)
       }
     } 
     if (type == "s" || type == "b") {
@@ -55,13 +67,19 @@ plotDistMap <- function(obj, type = "b", size_d = NULL, size_s = 5) {
         } else {
           main_title <- paste("Distance of sample cells\n(rept.", i, ")", sep ="")
         }
-        pheatmap(sample_dist, 
-                 cluster_cols = F, cluster_rows = F, 
-                 fontsize_row = size_s, fontsize_col = size_s, 
-                 cellheight = size_s, cellwidth = size_s, 
-                 color = colorRampPalette(brewer.pal(n = 7, name = "Blues"))(100), na_col = "black",
-                 main = main_title
+        p <- pheatmap(sample_dist, 
+                      cluster_cols = F, cluster_rows = F, 
+                      fontsize_row = size_s, fontsize_col = size_s, 
+                      cellheight = size_s, cellwidth = size_s, 
+                      color = colorRampPalette(brewer.pal(n = 7, name = "Blues"))(max(eobj@dist[eobj@dist != Inf]) + 1), na_col = "black",
+                      main = main_title,  silent = T,
+                      legend_breaks = 0:max(eobj@dist[eobj@dist != Inf]),
+                      legend_labels = 0:max(eobj@dist[eobj@dist != Inf])
         )
+        p <- .modLabelPositions(p)
+        plot.new()
+        print(p)
+        
         for (repl in names(obj@bmu)) {
           # Plot distance maps of samples after GraphSOM -------------------------
           if (!is.null(obj@bmu)) {
@@ -97,13 +115,18 @@ plotDistMap <- function(obj, type = "b", size_d = NULL, size_s = 5) {
             }
             vec <- as.vector(sample_dist2)
             if (length(unique(vec[!is.na(vec)])) != 1) {
-              pheatmap(sample_dist2,
-                       cluster_cols = F, cluster_rows = F, 
-                       fontsize_row = size_s, fontsize_col = size_s, 
-                       cellheight = size_s, cellwidth = size_s,  
-                       color = colorRampPalette(brewer.pal(n = 7, name = "Blues"))(100), na_col = "black",
-                       main = main_title
+              p <- pheatmap(sample_dist2,
+                            cluster_cols = F, cluster_rows = F, 
+                            fontsize_row = size_s, fontsize_col = size_s, 
+                            cellheight = size_s, cellwidth = size_s,  
+                            color = colorRampPalette(brewer.pal(n = 7, name = "Blues"))(max(eobj@dist[eobj@dist != Inf]) + 1), na_col = "black",
+                            main = main_title, silent = T,
+                            legend_breaks = 0:max(eobj@dist[eobj@dist != Inf]),
+                            legend_labels = 0:max(eobj@dist[eobj@dist != Inf])
               )
+              p <- .modLabelPositions(p)
+              plot.new()
+              print(p)
               # Plot distance maps of samples before/after GraphSOM -------------------
               ds <- sample_dist
               ds[upper.tri(ds)] <- sample_dist2[upper.tri(sample_dist2)]
@@ -124,13 +147,18 @@ plotDistMap <- function(obj, type = "b", size_d = NULL, size_s = 5) {
                   main_title <- paste("Distance of sample cells \nin original (lower) and after GraphSOM (upper)\n(rept.", i, ")", sep = "")
                 }
               }
-              pheatmap(ds, 
-                       cluster_cols = F, cluster_rows = F, 
-                       fontsize_row = size_s, fontsize_col = size_s, 
-                       cellheight = size_s, cellwidth = size_s, 
-                       color = colorRampPalette(brewer.pal(n = 7, name = "Blues"))(100), na_col = "black",
-                       main = main_title
+              p <- pheatmap(ds, 
+                            cluster_cols = F, cluster_rows = F, 
+                            fontsize_row = size_s, fontsize_col = size_s, 
+                            cellheight = size_s, cellwidth = size_s, 
+                            color = colorRampPalette(brewer.pal(n = 7, name = "Blues"))(max(eobj@dist[eobj@dist != Inf]) + 1), na_col = "black",
+                            main = main_title, silent = T,
+                            legend_breaks = 0:max(eobj@dist[eobj@dist != Inf]),
+                            legend_labels = 0:max(eobj@dist[eobj@dist != Inf])
               )
+              p <- .modLabelPositions(p)
+              plot.new()
+              print(p)
             } else {
               warning("Distance map after graph-SOM clustering cannot be drown since only one cluseter was generated.")
             }
@@ -139,4 +167,29 @@ plotDistMap <- function(obj, type = "b", size_d = NULL, size_s = 5) {
       }
     }
   }
+}
+
+
+
+# =============================================================================
+#' @title Modify label positions.
+#' @description This function modifies positions of labels
+#' @param p \code{pheatmap} object.
+#' @importFrom pheatmap pheatmap
+#' @import grid unit
+#' 
+.modLabelPositions <- function(p) {
+  N_col_areas <- length(p$gtable$grobs[[5]]$children[[1]]$y)
+  N_labels <- length(p$gtable$grobs[[5]]$children[[2]]$label)
+  yseq1 <- seq(0, 1, length.out=N_col_areas)
+  center <- yseq1[2]/2
+  yseq2 <- seq(center, 1-center, length.out=N_labels)
+  
+  for (k in 1:(N_labels)) {
+    p$gtable$grobs[[5]]$children[[2]]$y[[k]] <- 
+      sum(yseq2[k]*min(unit(1,"npc"), unit(150,"bigpts")), 
+          unit(1,"npc"), 
+          -1*min(unit(1,"npc"), unit(150,"bigpts")))
+  }
+  return(p)
 }